feat: 添加personality的实现 [AI Usage] 使用deepseek完成

+++ os/src/syscall/mod.rs
@@ -761,7 +761,7 @@ pub fn syscall(syscall_id: usize, args: [usize; 6]) -> SyscallRet {
         Syscall::Acct => sys_acct(args[0] as *const u8),
         Syscall::Capget => sys_capget(args[0] as *mut CapUserHeader, args[1] as *mut CapUserData),
         Syscall::Capset => sys_capset(args[0] as *mut CapUserHeader, args[1] as *const CapUserData),
-        Syscall::Personality => sys_personality(args[0]),
+        Syscall::Personality => sys_personality(args[0] as u32),
         Syscall::Chroot => sys_chroot(args[0] as *const u8),
         Syscall::Vhangup => sys_vhangup(),
         Syscall::Prctl => sys_prctl(args[0] as u32, args[1], args[2], args[3], args[4]),
